+++ lib/utils.ts
@@ -5,7 +5,9 @@ export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
 
-// This check can be removed, it is just for tutorial purposes
-export const hasEnvVars =
-  process.env.NEXT_PUBLIC_SUPABASE_URL &&
-  process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY;
+export function capitalize(text: string) {
+  if (typeof text !== "string" || text.length === 0) {
+    return "";
+  }
+  return text.charAt(0).toUpperCase() + text.slice(1);
+}